+++ client/src/features/common/ImageThumbnail.tsx
@@ -1,6 +1,6 @@
 /* eslint-disable react/jsx-props-no-spreading */
 import React, { useRef, useState } from 'react'
-import { Link } from 'react-router-dom'
+import { Link, useLocation } from 'react-router-dom'
 import sanitizeHtml from 'sanitize-html'
 import Button from 'react-bootstrap/Button'
 import Overlay from 'react-bootstrap/Overlay'
@@ -28,6 +28,7 @@ const popperConfig = {
 }
 
 const ImageThumbnail: React.FC<IProps> = ({ imageInfo, linkUrl }) => {
+  const { pathname } = useLocation()
   const [show, setShow] = useState(false)
   const attribRef = useRef(null)
 
@@ -47,6 +48,10 @@ const ImageThumbnail: React.FC<IProps> = ({ imageInfo, linkUrl }) => {
               'Results Snippet Link',
             )
           }
+          state={{
+            prevPath: pathname,
+            targetName: linkUrl,
+          }}
           data-testid="image-link"
         >
           <img