+++ client/src/features/common/TimelineContainers.tsx
@@ -2,7 +2,7 @@
 import React, { useState } from 'react'
 import { Col, Row } from 'react-bootstrap'
 import styled from 'styled-components'
-import { Link } from 'react-router-dom'
+import { Link, useLocation } from 'react-router-dom'
 
 import theme from '../../styles/theme'
 import { IHalLinks } from '../../types/IHalLinks'
@@ -59,6 +59,7 @@ const Relations: React.FC<{
   year: string
   searchTag: string
 }> = ({ searchTags, data, year, searchTag }) => {
+  const { pathname } = useLocation()
   const facetNameMap: Map<string, string> = new Map([
     ['itemProductionDate', 'Objects Produced'],
     ['itemEncounteredDate', 'Objects Encountered'],
@@ -68,6 +69,7 @@ const Relations: React.FC<{
 
   const { tab, jsonSearchTerm } = searchTags[searchTag]
   const { criteria, totalItems } = data[year][searchTag] as ITimelineCriteria
+  const searchQ = formatDateJsonSearch(year, jsonSearchTerm as string, criteria)
   return (
     <HoverableRow key={`${searchTag}-${year}`}>
       <Col xs={12} sm={12} md={6} lg={12} xl={6}>
@@ -80,11 +82,11 @@ const Relations: React.FC<{
           <Link
             to={{
               pathname: `/view/results/${tab}`,
-              search: `q=${formatDateJsonSearch(
-                year,
-                jsonSearchTerm as string,
-                criteria,
-              )}&collapseSearch=true`,
+              search: `q=${searchQ}&collapseSearch=true`,
+            }}
+            state={{
+              prevPath: pathname,
+              targetName: `/view/results/${tab}q=${searchQ}&collapseSearch=true`,
             }}
             onClick={() =>
               pushSiteImproveEvent('Search Link', 'Selected', 'Timeline')