+++ client/src/features/relatedLists/RelatedListSearchLink.tsx
@@ -1,6 +1,7 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import React from 'react'
 import { HashLink as Link } from 'react-router-hash-link'
+import { useLocation } from 'react-router-dom'
 
 import { scopeToTabTranslation } from '../../config/searchTypes'
 import { IAdvancedSearchState } from '../../redux/slices/advancedSearchSlice'
@@ -23,17 +24,23 @@ const RelatedListSearchLink: React.FC<ILinkParams> = ({
   total,
   label,
 }) => {
+  const { pathname } = useLocation()
   const tab = scopeToTabTranslation[scope]
 
   const linkLabel = `Show all ${total || ''} ${label || ''} result${
     total !== 1 ? 's' : ''
   }`
 
+  const searchQ = JSON.stringify(criteria)
   return (
     <Link
       to={{
         pathname: `/view/results/${tab}`,
-        search: `q=${JSON.stringify(criteria)}&openSearch=false`,
+        search: `q=${searchQ}&openSearch=false`,
+      }}
+      state={{
+        prevPath: pathname,
+        targetName: `/view/results/${tab}?q=${searchQ}&openSearch=false`,
       }}
       onClick={() =>
         pushSiteImproveEvent('Search Link', 'Selected', `Accordion ${title}`)