+++ client/src/features/search/SearchContainer.tsx
@@ -1,6 +1,7 @@
 import React, { useState } from 'react'
 import { Col, Row } from 'react-bootstrap'
 import { HashLink } from 'react-router-hash-link'
+import { useLocation } from 'react-router-dom'
 
 import { LinksContainerRow } from '../../styles/features/search/LinksContainerRow'
 import theme from '../../styles/theme'
@@ -32,6 +33,7 @@ const SearchContainer: React.FC<IProps> = ({
   },
   isResultsPage = false,
 }) => {
+  const { pathname, search } = useLocation()
   const [isError, setIsError] = useState<boolean>(false)
 
   return (
@@ -60,6 +62,10 @@ const SearchContainer: React.FC<IProps> = ({
             <Col xs={6} className="d-inline-flex justify-content-end">
               <HashLink
                 to="/content/simple-search"
+                state={{
+                  prevPath: `${pathname}${search}`,
+                  targetName: 'SearchTips',
+                }}
                 style={{
                   ...searchTipsStyle,
                   fontWeight: '400',