+++ client/CHANGELOG.md
@@ -57,6 +57,10 @@ All notable changes to this project will be documented in this file.
 - Added Place with Type on concept pages ([#107](https://github.com/project-lux/lux-frontend/issues/107)).
 - Added Events with Type on concept pages ([#128](https://github.com/project-lux/lux-frontend/issues/128)).
 - Added entity icon to About/Depicts ([#86](https://github.com/project-lux/lux-frontend/issues/86)).
+<<<<<<< HEAD
+=======
+- Added entity icon to About/Depicts ([#124](https://github.com/project-lux/lux-frontend/issues/124)).
+>>>>>>> 2eaec93 (124 added SiteImprove page tracking)
 
 ### Changed